Integrate Sentry only on production
