+++ portal/server/sentry.client.config.ts
@@ -7,12 +7,14 @@
 
 import * as Sentry from "@sentry/nextjs";
 
-Sentry.init({
-  dsn: process.env.SENTRY_DSN,
+if (process.env.NODE_ENV === "production") {
+    Sentry.init({
+        dsn: process.env.SENTRY_DSN,
 
-  // Define how likely traces are sampled. Adjust this value in production, or use tracesSampler for greater control.
-  tracesSampleRate: 1,
+        // Define how likely traces are sampled. Adjust this value in production, or use tracesSampler for greater control.
+        tracesSampleRate: 1,
 
-  // Setting this option to true will print useful information to the console while you're setting up Sentry.
-  debug: true,
-});
+        // Setting this option to true will print useful information to the console while you're setting up Sentry.
+        debug: true,
+    });
+}
